+++ app/helpers/media_magick/application_helper.rb
@@ -7,7 +7,7 @@ module MediaMagick
         relation: relation.to_s
       }
 
-      data_attributes.merge!(:partial => options[:partial]) if options[:partial]
+      data_attributes.merge!(:partial => get_partial_name(options))
       data_attributes.merge!(:embedded_in_id => options[:embedded_in].id.to_s, :embedded_in_model => options[:embedded_in].class.to_s) if options[:embedded_in]
 
       content_tag :div, id: model.class.to_s.downcase << '-' << relation.to_s, class: 'attachmentUploader ' << relation.to_s, data: data_attributes do
@@ -19,10 +19,24 @@ module MediaMagick
             relations:         relation,
             newAttachments:    options[:newAttachments]    || {},
             loadedAttachments: options[:loadedAttachments] || {},
-            partial:           options[:partial]
+            partial:           get_partial_name(options)
           }
 
           render '/upload', partial_attributes
+        end
+      end
+    end
+
+    private
+
+    def get_partial_name(options)
+      if options[:partial]
+        options[:partial]
+      else
+        if options[:as]
+          "/#{options[:as]}"
+        else
+          '/image'
         end
       end
     end