+++ spec/views/_upload.html.erb_spec.rb
@@ -3,10 +3,9 @@ require 'spec_helper'
 describe '_upload' do
   let(:album)             { stub_model(Album) }
   let(:album_with_photos) { stub_model(Album, photos: [stub_model(AlbumPhotos)]) }
-  let(:album_with_files)  { stub_model(Album, files: [stub_model(AlbumFiles)]) }
 
   def default_attributes
-    { newAttachments: {}, loadedAttachments: {}, partial: nil, model: album, relations: 'photos' }
+    { newAttachments: {}, loadedAttachments: {}, partial: '/image', model: album, relations: 'photos' }
   end
 
   it 'should personalize css classes attributes of newAttachments element' do
@@ -17,18 +16,6 @@ describe '_upload' do
   it 'should personalize css classes attributes of loadedAttachments element' do
     render 'upload', default_attributes.merge(loadedAttachments: { class: 'span3' })
     rendered.should =~ /<ul class="loadedAttachments span3">/
-  end
-
-  it 'should render image partial when attach is a image' do
-    stub_template '_image' => 'image partial'
-    render 'upload', default_attributes.merge(model: album_with_photos)
-    rendered.should =~ /image partial/
-  end
-
-  it 'should render image partial when attach is a file' do
-    stub_template '_file' => 'file partial'
-    render 'upload', default_attributes.merge(model: album_with_files, relations: 'files')
-    rendered.should =~ /file partial/
   end
 
   it 'should render the partial passed' do